requirements.txt
opencv-python-headless
Pillow==10.3.0
pytz==2024.1
SQLAlchemy==2.0.30
streamlit==1.35.0
ultralytics==8.2.2
streamlit-webrtc==0.47.7
